% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibInfoworks.R
\name{hsCreateGerrisInputConstParsCsv}
\alias{hsCreateGerrisInputConstParsCsv}
\title{Create Gerris Input File for Constant Parameters}
\usage{
hsCreateGerrisInputConstParsCsv(
  csvExample,
  csvOut = NULL,
  constParVals = hsGerrisConstParsMia(),
  dbg = TRUE
)
}
\arguments{
\item{csvExample}{path to example csv file containing names of Gerris boundary conditions
in first row, starting in second column (first column is timestamp)}

\item{csvOut}{full path to csv file to which non-simulated parameters are to be written.
If missing, this function returns a character matrix containing the
file content.}

\item{constParVals}{Data frame with columns \emph{VO2}, \emph{SI}, \emph{VPH}, \emph{VX0},
\emph{VX02}, \emph{ZOOIND}, \emph{VKIGR}, \emph{LF}, \emph{MW},
\emph{VNO2}, \emph{VNO3}, \emph{CA}, \emph{CHLA}, \emph{ANTBL} containing
the values for non-simulated parameters in the first and only row.
Default: result of hsGerrisConstParsMia().}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\description{
Create Gerris input file containing boundary conditions for non-simulated
  water quality parameters
}
