% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibInfoworks.R
\name{hsCreateGerrisInputFile}
\alias{hsCreateGerrisInputFile}
\title{Write File for Gerris Import}
\usage{
hsCreateGerrisInputFile(
  mdb,
  tbl,
  gerrisParID,
  csv.dir,
  csv.file = paste0("forGerrisImport_", gerrisParID, ".csv"),
  writeFile = TRUE,
  tsFormat = .defaultTimeFormat(),
  subst.na = "",
  dbg = FALSE
)
}
\arguments{
\item{mdb}{full path to MS Access database containing table with required
data}

\item{tbl}{name of table to be exported to csv file}

\item{gerrisParID}{Gerris parameter ID: one of "OBSB", "OCSB", "VNH4",
"GESN", "GELP", "GESP", "SS", "Q"}

\item{csv.dir}{path to directory to which output file is to be written}

\item{csv.file}{optional name for output file, default:
"forGerrisImport_<gerrisParID>.csv"}

\item{writeFile}{only if TRUE, a file is written, otherwise the corresponding
data is only returned by this function but not written to file.}

\item{tsFormat}{format of timestamp to be used in output file (\%d = day, \%m =
month, \%Y = year, \%H = hour, \%M = minute, \%S = second)}

\item{subst.na}{string value by which NULL values in table are subsubstituted
in the output file}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\description{
Write a file for Gerris import based on a database table generated by
  StatAnalysis-evaluation "IwToQSim".
}
