% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibInfoworks.R
\name{hsCreateGerrisInputFiles}
\alias{hsCreateGerrisInputFiles}
\title{Create Gerris Input Files}
\usage{
hsCreateGerrisInputFiles(
  mdb,
  csv.dir,
  tblQ = "tbl_Q_m3_s",
  tblBOD = "tbl_c_BOD_tot_mg_L",
  tblCOD = "tbl_c_COD_tot_mg_L",
  tblNH4N = "tbl_c_NH4N_mg_L",
  tblNges = "tbl_c_NGES_mg_L",
  tblTPdis = "tbl_c_TP_dis_mg_L",
  tblTPtot = "tbl_c_TP_tot_mg_L",
  tblTSS = "tbl_c_TSS_mg_L",
  csv.basename = "forGerrisImport",
  separate = TRUE,
  overall = TRUE,
  tsFormat = .defaultTimeFormat(),
  subst.na = "",
  dbg = FALSE
)
}
\arguments{
\item{mdb}{full path to MS Access database (.mdb) containing tables prepared
by StatAna-Evaluation "eIwToQSim"}

\item{csv.dir}{output directory to which database tables are to be exported
in CSV format}

\item{tblQ}{name of table containing flows Q in m3/s; default: "tbl_Q_m3_s"}

\item{tblBOD}{name of table containing BOD concentrations in mg/L; default:
"tbl_c_BOD_tot_mg_L"}

\item{tblCOD}{name of table containing COD concentrations in mg/L; default:
"tbl_c_COD_tot_mg_L"}

\item{tblNH4N}{name of table containing NH4-N concentrations in mg/L;
default: "tbl_c_NH4N_mg_L"}

\item{tblNges}{name of table containing N total concentrations in mg/L;
default: "tbl_c_NGES_mg_L"}

\item{tblTPdis}{name of table containing total P (dissolved) concentrations
in mg/L; default: "tbl_c_TP_dis_mg_L"}

\item{tblTPtot}{name of table containing total P concentrations in mg/L;
default: "tbl_c_TP_tot_mg_L"}

\item{tblTSS}{name of table containing TSS concentrations in mg/L; default:
"tbl_c_TSS_mg_L"}

\item{csv.basename}{basename of file(s) to be created in \emph{csv.dir};
default: "forGerrisImport". For flows, "_Hydrax" will be appended to the
basename, and for concentrations of water quality parameters "_QSim_<par>",
where <par> is one of "OBSB", "OCSB", "VNH4", "GESN", "GELP", "GESP", "SS".
To the file containing concentrations of all the water quality parameters
"_QSim_all" is appended.}

\item{separate}{if TRUE, one file per parameter is created.}

\item{overall}{if TRUE, one file containing all parameters is created.}

\item{tsFormat}{format of timestamp to be used in output file (\%d = day, \%m
= month, \%Y = year, \%H = hour, \%M = minute, \%S = second)}

\item{subst.na}{substitution value for NA values}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\description{
Create Gerris input files from database containing tables prepared by 
  StatAna-Evaluation "eIwToQSim"
}
