% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibInfoworks.R
\name{hsDaysInFile}
\alias{hsDaysInFile}
\title{Number of Days in InfoWorks Result CSV File}
\usage{
hsDaysInFile(bytesFile, bytesHeader, bytesRow, timestep)
}
\arguments{
\item{bytesFile}{file length in bytes}

\item{bytesHeader}{length of header line in bytes}

\item{bytesRow}{length of data row in bytes}

\item{timestep}{result \code{timestep} in seconds}
}
\value{
number of days \dQuote{contained} in an InfoWorks result CSV file of size
  \emph{bytesFile} with a header line of \emph{bytesHeader} bytes length, 
  each data row being \emph{bytesRow} bytes long and a result \code{timestep} of
  \emph{timestep} seconds.
}
\description{
Number of Days in InfoWorks Result CSV File
}
