% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibInfoworks.R
\name{hsFileSize}
\alias{hsFileSize}
\title{Size of InfoWorks Result CSV File}
\usage{
hsFileSize(nDays, bytesHeader, bytesRow, timestep, dbg = FALSE)
}
\arguments{
\item{nDays}{number of days}

\item{bytesHeader}{length of header line in bytes}

\item{bytesRow}{length of data row in bytes}

\item{timestep}{result \code{timestep} in seconds}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\value{
Size of an InfoWorks result CSV file over \emph{nDays} with a result
  \code{timestep} of \emph{timestep} seconds in bytes if the header file 
  is \emph{bytesHeader} and each data row is \emph{bytesRow} bytes long.
}
\description{
Size of an InfoWorks result CSV file in bytes
}
