% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibInfoworks.R
\name{hsGerrisConstBoundMatrix}
\alias{hsGerrisConstBoundMatrix}
\title{Gerris Constant Boundary Matrix}
\usage{
hsGerrisConstBoundMatrix(boundNames, constPars, tstamp.1, tstamp.2, tstamp.n)
}
\arguments{
\item{boundNames}{names of locations at which boundary conditions are given}

\item{constPars}{values of parameters that are treated as constant boundary conditions}

\item{tstamp.1}{first timestamp}

\item{tstamp.2}{second timestamp}

\item{tstamp.n}{last timestamp}
}
\description{
Character matrix containing default values for non-simulated parameters
}
