% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibInfoworks.R
\name{hsGerrisConstParsMia}
\alias{hsGerrisConstParsMia}
\title{Gerris Constant Parameters}
\usage{
hsGerrisConstParsMia(
  VO2 = 0,
  SI = 0,
  VPH = 7.44,
  VX0 = 0,
  VX02 = 0,
  ZOOIND = 0,
  VKIGR = 0.33,
  LF = 293.33,
  MW = 1.39,
  VNO2 = 0.21,
  VNO3 = 1.1,
  CA = 36.04,
  CHLA = 0,
  ANTBL = 0.33
)
}
\arguments{
\item{VO2}{"Sauerstoffgehalt" in mg/l}

\item{SI}{"Silizium" in mg/l}

\item{VPH}{"pH-Wert"}

\item{VX0}{"Nitrosomonas" in mg/l}

\item{VX02}{"Nitrobacter" in mg/l}

\item{ZOOIND}{"Rotatorien/Zooplanktondichte" in Ind/l}

\item{VKIGR}{"Kieselalgen/Anteil der Kieselalgen am Gesamt-Chlorophyll-a", 0..1}

\item{LF}{Leitfaehigkeit" in mikro-S/cm}

\item{MW}{"m-Wert" in mmol/l}

\item{VNO2}{"Nitrit-N" in mg/l}

\item{VNO3}{"Nitrat-N" in mg/l}

\item{CA}{"Calcium" in mg/l}

\item{CHLA}{"Chlorophyll-a" in mikro-g/l}

\item{ANTBL}{"Blaualgen/Anteil der Blaualgen am Gesamt-Chlorophyll-a", 0..1}
}
\value{
Data frame with columns \emph{VO2}, \emph{SI}, \emph{VPH}, \emph{VX0},
  \emph{VX02}, \emph{ZOOIND}, \emph{VKIGR}, \emph{LF}, \emph{MW},
  \emph{VNO2}, \emph{VNO3}, \emph{CA}, \emph{CHLA}, \emph{ANTBL} containing
  the default values for non-simulated parameters in the first and only row
}
\description{
Gerris/QSim boundary conditions that are treated as constant within MIA-CSO
  project. Parameter names according to definition in Gerris configuration
  file "GerrisParam.xml"
}
