% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibPlotIwResult.R
\name{hsGetIwResult}
\alias{hsGetIwResult}
\title{Get Infoworks Result}
\usage{
hsGetIwResult(file, columns = NULL, skip.columns = "Seconds", dbg = FALSE)
}
\arguments{
\item{file}{\code{file}}

\item{columns}{defalut: NULL}

\item{skip.columns}{default: "Seconds"}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\value{
data frame
}
\description{
Get Infoworks Result
}
