% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibPlotIwResult.R
\name{hsGetIwResultAvgAboveZero}
\alias{hsGetIwResultAvgAboveZero}
\title{Get Infoworks Result Average Above Zero}
\usage{
hsGetIwResultAvgAboveZero(src, type)
}
\arguments{
\item{src}{either full path to mdb or to  directory in which InfoWorks result files
(csv) are located}

\item{type}{one of "flow", "BOD_tot", "COD_tot", "NH4N", "TKN_tot", "TP_dis", 
"TP_tot", "TSS"}
}
\description{
Get Infoworks Result Average Above Zero
}
