% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibPlotIwResult.R
\name{hsGetIwResultAvgFromCsv}
\alias{hsGetIwResultAvgFromCsv}
\title{Get Infoworks Result Average From CSV}
\usage{
hsGetIwResultAvgFromCsv(csvdir, type, qthreshold = 0.003, dbg = TRUE)
}
\arguments{
\item{csvdir}{directory in which InfoWorks result files (csv) are located}

\item{type}{one of "flow", "BOD_tot", "COD_tot", "NH4N", "TKN_tot", "TP_dis", 
"TP_tot", "TSS"}

\item{qthreshold}{threshold for Q values}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\description{
Get Infoworks Result Average From CSV
}
