% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibPlotIwResult.R
\name{hsGetIwResultAvgFromMdb}
\alias{hsGetIwResultAvgFromMdb}
\title{Get Infoworks Result Average From MS Access Database}
\usage{
hsGetIwResultAvgFromMdb(
  mdb,
  type,
  tblQ = "tbl_05_Q_bei_Ueberlauf_m3_s",
  tblL = paste("tbl_02_15minMittel_L", type, "kg_s", sep = "_"),
  skipCols = c("^kug$", "^overall$")
)
}
\arguments{
\item{mdb}{\code{mdb}}

\item{type}{\code{type}}

\item{tblQ}{name of database table containing flows}

\item{tblL}{name of database table containing loads}

\item{skipCols}{vector of patterns matching names of columns to be skipped}
}
\description{
Get Infoworks Result Average From MS Access Database
}
