% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibPlotIwResult.R
\name{hsIntegrals}
\alias{hsIntegrals}
\title{Integrals}
\usage{
hsIntegrals(data, evts, dbg = FALSE)
}
\arguments{
\item{data}{time-series \code{data} with timestamp in first column}

\item{evts}{event \code{data} describing events contained in \emph{data}}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\value{
vector of integral values with one value per event. Length of vector
  corresponds to number of rows in \emph{evts}
}
\description{
Calculates the integrals (sum of all value columns within event's time 
  interval, multiplied with "signal width" of events)
}
