% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibPlotIwResult.R
\name{hsIwEventSummary}
\alias{hsIwEventSummary}
\title{Volume and Mass Load per Event}
\usage{
hsIwEventSummary(
  src,
  mfpars = c("flow", "BOD_tot", "COD_tot", "NH4N", "TKN_tot", "TP_dis", "TP_tot",
    "TSS")
)
}
\arguments{
\item{src}{Data source; can be either path to mdb (filled by StatAna) or
directory containing original csv files exported from InfoWorks}

\item{mfpars}{Vector of parameter acronyms; Default: c("flow", "BOD_tot",
"COD_tot", "NH4N", "TKN_tot", "TP_dis", "TP_tot", "TSS")}
}
\value{
list with event list \emph{evts}, matrix \emph{allIntegrals}
  containing volume and mass loads per event and data.frame \emph{iwdata}
  containing all result data
}
\description{
Volume and Mass Load per Event
}
