% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibPlotIwResult.R
\name{hsIwParNameMap}
\alias{hsIwParNameMap}
\title{Parameter Name Mapping}
\usage{
hsIwParNameMap()
}
\description{
Mapping between parameter names used in table names in temporary mdb
  (written to by StatAna) and parameter acronyms used in InfoWorks result 
  csv files.
}
