% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibPlotIwResult.R
\name{hsIwPlot1}
\alias{hsIwPlot1}
\title{Infoworks Plot 1}
\usage{
hsIwPlot1(
  data,
  allIntegrals,
  evts,
  type = 2,
  basemain,
  legend.sort = FALSE,
  ylog = FALSE,
  yBottom = ifelse(ylog, 1e-04, 0),
  cex.legend = 0.55
)
}
\arguments{
\item{data}{\code{data}}

\item{allIntegrals}{\code{allIntegrals}}

\item{evts}{\code{evts}}

\item{type}{default: 2}

\item{basemain}{\code{basemain}}

\item{legend.sort}{default: FALSE}

\item{ylog}{default: FALSE}

\item{yBottom}{default: ifelse(\code{ylog}, 0.0001, 0)}

\item{cex.legend}{expansion factor for legend texts}
}
\description{
Infoworks Plot 1
}
