% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibPlotIwResult.R
\name{hsIwPlot2}
\alias{hsIwPlot2}
\title{Infoworks Plot 2}
\usage{
hsIwPlot2(
  allIntegrals,
  evts,
  plotTotal = TRUE,
  plotEvents = FALSE,
  pars = c("TSS", "COD_tot", "BOD_tot", "TKN_tot", "NH4N", "TP_tot", "TP_dis")
)
}
\arguments{
\item{allIntegrals}{\code{allIntegrals}}

\item{evts}{\code{evts}}

\item{plotTotal}{default: TRUE}

\item{plotEvents}{default: FALSE}

\item{pars}{default: c("TSS", "COD_tot", "BOD_tot", "TKN_tot", "NH4N", "TP_tot", "TP_dis")}
}
\description{
Infoworks Plot 2
}
