% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibPlotIwResult.R
\name{hsIwPlot3}
\alias{hsIwPlot3}
\title{Infoworks Plot 3}
\usage{
hsIwPlot3(allIntegrals, boxplot.range = 0)
}
\arguments{
\item{allIntegrals}{\code{allIntegrals}}

\item{boxplot.range}{this determines how far the plot whiskers extend out
from the box. See argument \code{range} of \code{\link{barplot}}.}
}
\description{
Infoworks Plot 3
}
