% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibPlotIwResult.R
\name{hsIwPlotAll}
\alias{hsIwPlotAll}
\title{Infoworks Plot All}
\usage{
hsIwPlotAll(
  src,
  subtitle = "",
  mfpars = c("flow", "BOD_tot", "COD_tot", "NH4N", "TKN_tot", "TP_dis", "TP_tot",
    "TSS"),
  ylog = FALSE,
  type = 2,
  outpdf = "",
  outdir = "",
  legend.sort = FALSE,
  cex.legend = 0.55
)
}
\arguments{
\item{src}{Data source; can be either path to mdb (filled by StatAna) or
directory containing original csv files exported from InfoWorks}

\item{subtitle}{\code{subtitle}}

\item{mfpars}{Vector of parameter acronyms; Default: c("flow", "BOD_tot",
"COD_tot", "NH4N", "TKN_tot", "TP_dis", "TP_tot", "TSS")}

\item{ylog}{y axis logarithmic?}

\item{type}{default: 2}

\item{outpdf}{Path to output file (.pdf). Default: ""}

\item{outdir}{Path to output directory. Default: ""}

\item{legend.sort}{default: FALSE}

\item{cex.legend}{expansion factor for legend texts}
}
\description{
Infoworks Plot All
}
