% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibInfoworks.R
\name{hsIwResultFileSize}
\alias{hsIwResultFileSize}
\title{InfoWorks Result CSV File Size}
\usage{
hsIwResultFileSize(
  dateFirst,
  dateLast,
  timestep,
  nDataCol,
  bytesHeader = -1,
  tstamp = "yyyy-mm-dd hh:nn:ss",
  colWidth = 12,
  dbg = FALSE
)
}
\arguments{
\item{dateFirst}{first date (day) to be simulated in ISO-format: \emph{yyyy-mm-dd}}

\item{dateLast}{last date (day) to be simulated in ISO-format: \emph{yyyy-mm-dd}}

\item{timestep}{result \code{timestep} in seconds}

\item{nDataCol}{number of data columns (time-columns excluded) in the InfoWorks result CSV file}

\item{bytesHeader}{length of header line in bytes}

\item{tstamp}{string representing an example timestamp}

\item{colWidth}{width of a data column in bytes}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\value{
List with elements \emph{Bytes}, \emph{kB}, \emph{MB}, \emph{GB}
  giving the requested file size in the according unit.
}
\description{
Size of an InfoWorks result CSV file depending on the simulated time period
  between \emph{dateFirst} and \emph{dateLast}, the result \code{timestep} 
  \emph{timestep} applied and the number \emph{nDataCol} of data columns in 
  the file.
}
