% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibPlotIwResult.R
\name{hsNameFilter}
\alias{hsNameFilter}
\title{Name Filter}
\usage{
hsNameFilter(x, posPtrn = NULL, negPtrn = NULL)
}
\arguments{
\item{x}{vector of character}

\item{posPtrn}{pattern matching elements to be included}

\item{negPtrn}{pattern matching elements to be excluded}
}
\value{
elements of \emph{x} matching \emph{posPtrn} and not matching \emph{negPtrn}
}
\description{
Include / exclude elements from vector by patterns
}
