% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibInfoworks.R
\name{hsPlotIwFileSizeVsTsAndPeriod}
\alias{hsPlotIwFileSizeVsTsAndPeriod}
\title{IW File Size vs. Time Step, Period}
\usage{
hsPlotIwFileSizeVsTsAndPeriod(
  nCols,
  colWidth = 12,
  tsFormat = "01.01.2011 00:00:00",
  bytesHeader = -1
)
}
\arguments{
\item{nCols}{number of data columns in the InfoWorks result CSV file}

\item{colWidth}{width of a data column in bytes}

\item{tsFormat}{string representing a timestamp}

\item{bytesHeader}{length of header line in bytes}
}
\description{
Plot showing InfoWorks result CSV file size vs. different combinations of
  timestep and simulated time period
}
