% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibInfoworks.R
\name{hsRep}
\alias{hsRep}
\title{Repeat \code{elements} n-times}
\usage{
hsRep(elements, n)
}
\arguments{
\item{elements}{vector of which each element is to be repeated n-times}

\item{n}{number by which each element of \emph{elements} is repeated}
}
\value{
vector in which each element of \emph{elements} is repeated \emph{n}-times
}
\description{
Repeat \code{elements} n-times
}
