% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibInfoworks.R
\name{hsRowLen}
\alias{hsRowLen}
\title{Row Length in InfoWorks Result CSV File}
\usage{
hsRowLen(colWidth, colNum, tsFormat = "2011-12-31 23:59:59")
}
\arguments{
\item{colWidth}{width of a data column in bytes}

\item{colNum}{number of data columns}

\item{tsFormat}{string representing a timestamp}
}
\value{
Number of bytes needed for one row of an InfoWorks result CSV file with
  \emph{colNum} data columns of \emph{colWidth} bytes each and a timestamp
  column of format according to the example timestamp \emph{tstamp}.
}
\description{
length of a data row in the InfoWorks result CSV file in bytes
}
