% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibPlotIwResult.R
\name{hsTimeBarPlot}
\alias{hsTimeBarPlot}
\title{Plot Type 1}
\usage{
hsTimeBarPlot(
  t1,
  t2,
  height,
  tlim = c(min(t1), max(t2)),
  ylim = NULL,
  ylog = TRUE,
  ymult = c(1, 2, 3, 5),
  main = "hsTimeBarPlot",
  tlab = "time",
  ylab = "height",
  time.format = .defaultDateFormat(),
  tlab.mindist = 86400,
  col = "grey",
  cex.all = 0.8,
  cex.legend = cex.all,
  cex.text = cex.all,
  cex.axis = cex.all,
  legend.values = height,
  legend.sort = TRUE,
  legend.format = .defaultLegendFormat(),
  legend.title = "Legend:",
  mar = c(6, 5, 5, 10)
)
}
\arguments{
\item{t1}{begin times of bars}

\item{t2}{end times of bars}

\item{height}{bar heights}

\item{tlim}{limits of time axis. Default: c(min(\code{t1}), max(\code{t2}))}

\item{ylim}{limits of y axis, Default: c(min(\code{height}),
max(\code{height}))}

\item{ylog}{if TRUE, y axis is scaled logarithmically, else linearly}

\item{ymult}{numbers to be multiplied by exponents of ten for labelling of y
axis when \emph{ylog} is TRUE}

\item{main}{plot title}

\item{tlab}{label of time axis}

\item{ylab}{label of y axis}

\item{time.format}{format of time axis labels}

\item{tlab.mindist}{minimum "time distance" between time labels, in seconds}

\item{col}{bar colour}

\item{cex.all}{general expansion factor to be applied to \code{cex.legend},
\code{cex.text}, \code{cex.axis}, if not given}

\item{cex.legend}{expansion factor for legend texts}

\item{cex.text}{expansion factor of texts within plot (bar numbers)}

\item{cex.axis}{expansion factor of axis labels}

\item{legend.values}{values to be shown in legend, default: \code{height}
values}

\item{legend.sort}{if TRUE, legend is sorted decreasingly by
\code{legend.values}}

\item{legend.format}{format string to be used by sprintf for generation of
legend entries from legend values, e.g. "V = \%8.0f m3"}

\item{legend.title}{legend title}

\item{mar}{plot margins}
}
\description{
Plot type 1: boxes with mean flow as \code{height} and event duration as
width
}
