% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibInfoworks.R
\name{hsTsInFile}
\alias{hsTsInFile}
\title{Result Time Step in InfoWorks Result CSV File}
\usage{
hsTsInFile(bytesFile, bytesHeader, bytesRow, nDays)
}
\arguments{
\item{bytesFile}{file length in bytes}

\item{bytesHeader}{length of header line in bytes}

\item{bytesRow}{lengh of data row in bytes}

\item{nDays}{number of days}
}
\value{
Returns the (possibly) applied result timestep of an InfoWorks simulation 
  depending on the maximal allowed InfoWorks result CSV file size \emph{bytesFile}
  in bytes, on the number \emph{nDays} of days to simulate and on the size
  \emph{bytesHeader} and \emph{bytesRow} of the header and of a data
  line, respectively.
}
\description{
Result Time Step in InfoWorks Result CSV File
}
